+++ runtime/t0rn-parachain/src/circuit_config.rs
@@ -1,6 +1,6 @@
 use crate::{
-    AccountId, AccountManager, Attesters, Balance, Balances, BlockNumber, Call, Circuit, Clock,
-    Event, Portal, RandomnessCollectiveFlip, Runtime, Timestamp, XDNS,
+    AccountId, AccountManager, Attesters, Aura, Balance, Balances, BlockNumber, Call, Circuit,
+    Clock, Event, Portal, RandomnessCollectiveFlip, Runtime, Timestamp, XDNS,
 };
 use sp_runtime::Percent;
 use sp_std::marker::PhantomData;
@@ -147,10 +147,12 @@ impl pallet_attesters::Config for Runtime {
     type MaxBatchSize = ConstU32<128>;
     type MinAttesterBond = MinAttesterBond;
     type MinNominatorBond = MinNominatorBond;
+    type Portal = Portal;
     type RandomnessSource = RandomnessCollectiveFlip;
     type RewardMultiplier = RewardMultiplier;
     type ShufflingFrequency = ConstU32<400>;
     type SlashAccount = EscrowAccount;
+    type Xdns = XDNS;
 }
 
 use t3rn_primitives::monetary::TRN;
@@ -182,6 +184,7 @@ impl pallet_rewards::Config for Runtime {
     type Event = Event;
     type ExecutorBootstrapRewards = ExecutorBootstrapRewards;
     type ExecutorInflation = ExecutorInflation;
+    type FindAuthor = pallet_session::FindAccountFromAuthorIndex<Self, Aura>;
     type InflationDistributionPeriod = InflationDistributionPeriod;
     type OneYear = OneYear;
     type TotalInflation = TotalInflation;