+++ CHANGELOG.md
@@ -4,6 +4,36 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.27.1-rc.0] - 2023-05-22
+### :bug: Bug Fixes
+- [`b05216d`](https://github.com/t3rn/t3rn/commit/b05216d82696740d946ce4a1d931580a808b717f) - grandpa range edge case *(PR [#1001](https://github.com/t3rn/t3rn/pull/1001) by [@petscheit](https://github.com/petscheit))*
+
+
+## [v1.27.0-rc.0] - 2023-05-19
+### :sparkles: New Features
+- [`5243ee1`](https://github.com/t3rn/t3rn/commit/5243ee12816ad84e990d19198d1ea1898b13725e) - batch attestations per target for committee transitions, penalties and SFX *(PR [#980](https://github.com/t3rn/t3rn/pull/980) by [@MaciejBaj](https://github.com/MaciejBaj))*
+
+
+## [v1.26.0-rc.0] - 2023-05-19
+### :recycle: Refactors
+- [`e991c25`](https://github.com/t3rn/t3rn/commit/e991c25737a597ae2573d32d3697751e10ad578b) - connect and use ReadSFX interface from Circuit to Attesters *(PR [#884](https://github.com/t3rn/t3rn/pull/884) by [@MaciejBaj](https://github.com/MaciejBaj))*
+
+
+## [v1.25.1-rc.0] - 2023-05-18
+### :bug: Bug Fixes
+- [`2825cd5`](https://github.com/t3rn/t3rn/commit/2825cd5b37483c17d5bbebfbe66f6d91ea7f4709) - add correct filter *(PR [#989](https://github.com/t3rn/t3rn/pull/989) by [@palozano](https://github.com/palozano))*
+
+
+## [v1.25.0-rc.0] - 2023-05-18
+### :sparkles: New Features
+- [`f45fa58`](https://github.com/t3rn/t3rn/commit/f45fa580e3a44b0a3a88c36ddb8a4e801eb137a4) - distribute inflation rewards to block authors  *(PR [#966](https://github.com/t3rn/t3rn/pull/966) by [@MaciejBaj](https://github.com/MaciejBaj))*
+
+
+## [v1.24.0-rc.0] - 2023-05-18
+### :recycle: Refactors
+- [`1311cd9`](https://github.com/t3rn/t3rn/commit/1311cd9d077d76a26447ddff7e511c8841818370) - sfx actions *(PR [#981](https://github.com/t3rn/t3rn/pull/981) by [@ahkohd](https://github.com/ahkohd))*
+
+
 ## [v1.23.0-rc.0] - 2023-05-17
 ### :sparkles: New Features
 - [`6991de8`](https://github.com/t3rn/t3rn/commit/6991de886f51a092a48ad1c562fb2d723e2e13b6) - add support for enforce executioner field *(PR [#970](https://github.com/t3rn/t3rn/pull/970) by [@ahkohd](https://github.com/ahkohd))*
@@ -1133,3 +1163,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v1.22.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.21.0-rc.0...v1.22.0-rc.0
 [v1.22.3-rc.0]: https://github.com/t3rn/t3rn/compare/v1.22.2-rc.0...v1.22.3-rc.0
 [v1.23.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.22.3-rc.0...v1.23.0-rc.0
+[v1.24.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.23.0-rc.0...v1.24.0-rc.0
+[v1.25.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.24.1-rc.0...v1.25.0-rc.0
+[v1.25.1-rc.0]: https://github.com/t3rn/t3rn/compare/v1.25.0-rc.0...v1.25.1-rc.0
+[v1.26.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.25.1-rc.0...v1.26.0-rc.0
+[v1.27.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.26.0-rc.0...v1.27.0-rc.0
+[v1.27.1-rc.0]: https://github.com/t3rn/t3rn/compare/v1.27.0-rc.0...v1.27.1-rc.0